[not for land] fix fx quant for quant_layer -> stack -> sum
Summary:

Before this PR, code patterns like this did not work:

```
x = some_quant_layer(x)
x = torch.stack([x, ...])
x = torch.sum(x, ...)
```

The reason this did not work is because `torch.sum` is treated as
"quantized" because of the newly added fp16 support, even though it is
not actually "quantized" for models where fp16 is not used.  We may
need to adjust the concept of "quantized vs non-quantized" into a
"dtype" for the longer term fix.

The current PR is a hacky fix to unblock.  We need to clean things
up before this is landable

Test Plan:

```
python test/test_quantization.py TestQuantizeFx.test_quant_sum
```

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 9a1894a69acc4c5de3f966b0a8a3ba0247764e44
Pull Request resolved: https://github.com/pytorch/pytorch/pull/53196