[package] Add `deny` method to `PackageExporter`
**Summary**
This commit adds a `deny` method to `PackageExporter` that allows
modules to be prohibited during the packaging process. A dependency on a
module matching the names or globs that `deny` was called with will
cause an exception to be raised.

**Test Plan**
This commit adds unit tests to `PackagingTest` for this new method:
`test_deny` and `test_deny_glob`.

**Fixes**
This commit fixes #53217.

ghstack-source-id: bdc7c47e36e7f2bae4adb980d85a77e0e74b5cf8
Pull Request resolved: https://github.com/pytorch/pytorch/pull/53233