[ONNX] Remove the last Cast in pow symbolic_opset9 (#52646)
Fixes #52436
For opset 9 of onnx Pow, if X is int32, Y is float, we will cast back to int32 which is consistent with X type.
However, pytorch is still float. The aten graph sometimes does not bind with the type for operators,
we are fine with the float type and don't want to cast back.
Even if X, Y are int32, the resulting float32 and int32 makes no difference.

ghstack-source-id: 25ce3d0f606dbc61a57603bbfc81614e2a0a5c8b
Pull Request resolved: https://github.com/pytorch/pytorch/pull/53305